% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bref_daily_stats.R
\name{bref_daily_Performance}
\alias{bref_daily_Performance}
\title{Scrape Batter/Pitcher Performance Data Over a Custom Time Frame}
\usage{
bref_daily_Performance(start_date, end_date, pit_bat)
}
\arguments{
\item{start_date}{First date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}

\item{end_date}{Last date data should be scraped from. Should take the form "YEAR-MONTH-DAY"}

\item{pit_bat}{}
}
\description{
This function allows you to scrape basic batter statistics over a custom time frame. Data is sourced from Baseball-Reference.com.
}
\examples{
\dontrun{bref_daily_Performance("2015-05-10", "2015-06-20", "pitcher")}
}
