% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_mlb_gamelog.R
\name{fg_batter_gamelog}
\alias{fg_batter_gamelog}
\title{Scrape Batter Game Logs from FanGraphs}
\usage{
fg_batter_gamelog(playerid, year)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}
}
\description{
This function allows you to scrape game logs by year for a batter from FanGraphs.com.
}
\examples{
\dontrun{fg_batter_gamelog(playerid = 6184, year = 2017)}
}
