% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_gamelog.R
\name{fg_milb_batter_gamelog}
\alias{fg_milb_batter_gamelog}
\title{# Scrape MiLB game logs for batters from Fangraphs, combining 'standard' and 'advanced' tabs}
\usage{
fg_milb_batter_gamelog(playerid, year)
}
\arguments{
\item{playerid}{The batter's minor leauge ID from FanGraphs.com.}

\item{year}{The season for which game logs should be returned.}
}
\description{
This function allows you to scrape MILB game logs for individual batters from FanGraphs.com.
}
\examples{
\dontrun{fg_milb_batter_gamelog(playerid = "sa917940", 2018)}
}
