% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_gamelog.R
\name{fg_milb_gamelog}
\alias{fg_milb_gamelog}
\title{# Scrape MiLB game logs for batters from Fangraphs, combining 'standard' and 'advanced' tabs}
\usage{
fg_milb_gamelog(playerid, year, pit_bat)
}
\arguments{
\item{playerid}{The batter's minor leauge ID from FanGraphs.com.}

\item{year}{The season for which game logs should be returned.}

\item{pit_bat}{pitcher or batter}
}
\description{
This function allows you to scrape MILB game logs for individual batters from FanGraphs.com.
}
\examples{
\dontrun{fg_milb_gamelog(playerid = 20126, 2018, "batter")}
}
