% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_milb_gamelog.R
\name{fg_milb_pitcher_gamelog}
\alias{fg_milb_pitcher_gamelog}
\title{# Scrape MiLB game logs for pitchers from Fangraphs, combining 'standard' and 'advanced' tabs}
\usage{
fg_milb_pitcher_gamelog(playerid, year)
}
\arguments{
\item{playerid}{The pitcher's minor leauge ID from FanGraphs.com.}

\item{year}{The season for which game logs should be returned.}
}
\description{
This function allows you to scrape MILB game logs for individual batters from FanGraphs.com.
}
\examples{
\dontrun{milb_pitcher_game_logs_fg(playerid = "sa873980", 2018)}
}
