% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_mlb_gamelog.R
\name{fg_mlb_gamelog}
\alias{fg_mlb_gamelog}
\title{Scrape Pitcher Game Logs from FanGraphs}
\usage{
fg_mlb_gamelog(playerid, year, pit_bat)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}

\item{pit_bat}{pitcher or batter}
}
\description{
This function allows you to scrape game logs by year for a pitcher from FanGraphs.com.
}
\examples{
\dontrun{fg_mlb_gamelog(playerid = 104, year = 2006, "pitcher")}
}
