% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_mlb_leaders.R
\name{fg_mlb_leaders}
\alias{fg_mlb_leaders}
\title{Scrape Leaderboards from FanGraphs}
\usage{
fg_mlb_leaders(year, league = "all", qual = "y", ind = 0, pit_bat)
}
\arguments{
\item{year}{Season for which you want data.}

\item{league}{Option for limiting results to different leagues or overall results. Options are "al", "nl", or "all".}

\item{qual}{Whether you want only batters/pitchers that qualified in a given season, or the minimum number of plate appearances for inclusion. If you only want qualified hitters, use qual. If a minimumm number of plate appearaces/innings pitched, use the number desired. Defaults to "y".}

\item{ind}{Whether or not to break the seasons out individual, or roll them up together. 1 = split seasons, 0 = aggregate seasons.}

\item{pit_bat}{pitcher or batter}
}
\description{
This function allows you to scrape all leaderboard statistics (basic and advanced) from FanGraphs.com.
}
\examples{
\dontrun{fg_mlb_leaders(year = 2015, pit_bat = "batter")}
}
