% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_mlb_gamelog.R
\name{fg_pitcher_gamelog}
\alias{fg_pitcher_gamelog}
\title{Scrape Pitcher Game Logs from FanGraphs}
\usage{
fg_pitcher_gamelog(playerid, year)
}
\arguments{
\item{playerid}{This is the playerid used by FanGraphs for a given player}

\item{year}{The season for which game logs should be returned (use the YYYY format)}
}
\description{
This function allows you to scrape game logs by year for a pitcher from FanGraphs.com.
}
\examples{
\dontrun{fg_pitcher_gamelog(playerid = 104, year = 2006)}
}
