% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{game_ids}
\alias{game_ids}
\title{A data set of game id}
\format{A vector with 35,180 elements.}
\usage{
data(game_ids)
}
\description{
A data set of game id
}
\details{
Dataset
}
\keyword{datasets}
