% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strikezones.R
\name{get_strikezones}
\alias{get_strikezones}
\title{Calculate strikezone boundaries}
\usage{
get_strikezones(data)
}
\arguments{
\item{data}{The data of scrape_statcast function.}
}
\value{
Returns a list of boundaries for both right handed batters and left handed batters
}
\description{
Calculate strikezone boundaries
}
