% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_ids}
\alias{player_ids}
\title{A data set of player id}
\format{A data frame with 2,040 rows and 12 variables}
\usage{
data(player_ids)
}
\description{
A data set of player id

\itemize{
  \item \code{IDPLAYER}: plyer id (maybe you dont use)
  \item \code{PLAYERNAME}: player name
  \item \code{BIRTHDATE}: birth date
  \item \code{FIRSTNAME}: first name
  \item \code{LASTNAME}: last name
  \item \code{TEAM}: team
  \item \code{LG}: league
  \item \code{POS}: position
  \item \code{IDFANGRAPHS}: plyer id in fangraphs
  \item \code{FANGRAPHSNAME}: plyer name in fangraphs
  \item \code{MLBID}: plyer id in statcast
  \item \code{MLBNAME}: plyer name in statcast
}
}
\details{
Dataset
}
\keyword{datasets}
