% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_run_expectancy.R
\name{run_expectancy_table}
\alias{run_expectancy_table}
\title{Generate run expectancy tables from Baseball Savant data}
\usage{
run_expectancy_table(df, level)
}
\arguments{
\item{df}{A data frame generated from Baseball Savant that has been formatted using the run_expectancy_code() function.}

\item{level}{Whether you want run expectancy calculated at the plate appearance or pitch level. "PA" or "Pitch"}
}
\description{
These functions allow a user to generate run expectancy tables from Baseball Savant data.
Tables are automatically assigned to the Global Environment.
}
\examples{
\dontrun{run_expectancy_table(df, level = "PA")}
}
