% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_batting_stats.R
\name{sc_batting_stats}
\alias{sc_batting_stats}
\title{Create statlines that include count and rate metrics for players based on Statcast}
\usage{
sc_batting_stats(df, base = "pa")
}
\arguments{
\item{df}{A data frame of statistics that includes, at a minimum, the following columns: events, description, game_date, and type.}

\item{base}{Tells the function what to use as the population of pitches to use for the statline. Options include "swings", "contact", or "pa". Defaults to "pa".}
}
\description{
This function allows you to create statlines of statistics for players or groups of players from raw Statcast. When calculating wOBA, the most recent year in the data frame is used for weighting.
}
\examples{
\dontrun{
sc_batting_stats(df, base = "contact")
}
}
