% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_linear_weights.R
\name{sc_linear_weights}
\alias{sc_linear_weights}
\title{Generate linear weight values for events using Baseball Savant data}
\usage{
sc_linear_weights(df, level)
}
\arguments{
\item{df}{A data frame generated from Baseball Savant that has been run through
the \code{\link{run_expectancy_code}} function.}

\item{level}{Whether to calculate linear weights the plate appearance or pitch level. "PA" or "Pitch"}
}
\description{
This function allows a user to generate linear weight values for events using Baseball Savant data. Output includes both linear weights above average and linear weights above outs for home runs, triples, doubles, singles, walks, hit by pitches, and outs.
}
\examples{
\dontrun{sc_linear_weights(df, level = "PA")}
}
