% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pbp.R
\name{sc_pbp}
\alias{sc_pbp}
\title{Scrape statcast data}
\usage{
sc_pbp(start_date, end_date)
}
\arguments{
\item{start_date}{First game date. Format must be in yyyy-mm-dd.}

\item{end_date}{Last game date. Format must be in yyyy-mm-dd.}
}
\description{
Scrape statcast data
}
\examples{
\dontrun{
sc_pbp(start_date = "2018-04-06", end_date = "2018-04-15")
}
}
