% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_snapshots.R
\name{sc_snapshots}
\alias{sc_snapshots}
\title{Produce time sequenced pitch locations from statcast parameters}
\usage{
sc_snapshots(df, interval = 0.01)
}
\arguments{
\item{df}{The data which have parameters used to determine the location of a pitch at a given time.}

\item{interval}{the amount of time between 'snapshots'}
}
\value{
Return a three dimensional array.
}
\description{
Produce time sequenced pitch locations from statcast parameters
}
\references{
http://baseball.physics.illinois.edu/KaganPitchfx.pdf
}
