% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_strikezone_mean.R
\name{sc_strikezone_mean}
\alias{sc_strikezone_mean}
\title{Calculate strikezone boundaries}
\usage{
sc_strikezone_mean(df)
}
\arguments{
\item{df}{The data of scrape_statcast function.}
}
\value{
Returns a list of boundaries for both right handed batters and left handed batters
}
\description{
Calculate strikezone boundaries
}
