% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsapi_batting_orders.R
\name{statsapi_batting_orders}
\alias{statsapi_batting_orders}
\title{Retrieve batting orders for a given MLB game via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
statsapi_batting_orders(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}

\item{type}{Whether to just return the starting lineup ('starting') or all
batters that appeared ('all')}
}
\value{
Returns a data frame that includes probable starting pitchers and
the home plate umpire for the game_pk provided
requested
}
\description{
Retrieve batting orders for a given MLB game via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
\dontrun{statsapi_batting_orders(566001)}
}
