% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsapi_gameinfo.R
\name{statsapi_gameinfo}
\alias{statsapi_gameinfo}
\title{Find game_pk values for professional baseball games (major and minor leagues)
via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
statsapi_gameinfo(year, level_ids = c(1, 11:17, 5442))
}
\arguments{
\item{year}{The year for which you want to find game_pk values for MLB games}

\item{level_ids}{A numeric vector with ids for each level where game_pks are
desired. See below for a reference of level ids.}
}
\value{
Returns a data frame that includes game_pk values and additional
information for games scheduled or played
}
\description{
Find game_pk values for professional baseball games (major and minor leagues)
via the MLB api \url{http://statsapi.mlb.com/api/}
}
\section{Level IDs}{


The following IDs can be passed to the level_ids argument:

1 = MLB
11 = Triple-A
12 = Doubl-A
13 = Class A Advanced
14 = Class A
15 = Class A Short Season
5442 = Rookie Advanced
16 = Rookie
17 = Winter League
}

\examples{
\dontrun{statsapi_gameinfo(2018:2019, 1))}
}
