% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsapi_gameinfo_pk.R
\name{statsapi_gameinfo_pk}
\alias{statsapi_gameinfo_pk}
\title{Retrieve additional game information for major and minor league games via the MLB api \url{http://statsapi.mlb.com/api/}}
\usage{
statsapi_gameinfo_pk(game_pk)
}
\arguments{
\item{game_pk}{The unique game_pk identifier for the game}
}
\value{
Returns a data frame that includes supplemental information, such as
weather, official scorer, attendance, etc., for the game_pk provided
}
\description{
Retrieve additional game information for major and minor league games via the MLB api \url{http://statsapi.mlb.com/api/}
}
\examples{
\dontrun{statsapi_gameinfo_pk(566001)}
}
