% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_results_ref.R
\name{team_results_ref}
\alias{team_results_ref}
\title{Scrape Team Results}
\usage{
team_results_ref(Tm, year)
}
\arguments{
\item{Tm}{The abbreviation used by Baseball-Reference.com for the team whose results you want to scrape.}

\item{year}{Season for which you want to scrape the park factors.}
}
\description{
This function allows you to scrape schedule and results for a major league team from Baseball-Reference.com
}
\examples{
team_results_ref("NYM", 2015)
team_results_ref("TBR", 2008)
}
