% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gameids.R
\name{get_gameids}
\alias{get_gameids}
\title{Get game_ids in specified years}
\usage{
get_gameids(start_year = 2018L, end_year = 2018L)
}
\arguments{
\item{start_year}{integer. Over 2008.}

\item{end_year}{integer.}
}
\value{
game_id vector
}
\description{
Get game_ids in specified years
}
