% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_statcast.R
\name{scrape_statcast}
\alias{scrape_statcast}
\title{Scrape statcast data}
\usage{
scrape_statcast(start_date, end_date)
}
\arguments{
\item{start_date}{Date of first game you want. Format must be in YYYY-MM-DD format.}

\item{end_date}{Date of last game you want. Format must be in YYYY-MM-DD format.}
}
\description{
Scrape statcast data
}
\examples{
\dontrun{
scrape_statcast(start_date = "2018-04-06", end_date = "2018-04-15")
scrape_statcast(start_date = "2018-04-06", end_date = "2018-04-15")
}
}
